FIX: 가구 조회 시 카테고리 반영 해결 (Integer -> Long)

See merge request s09-final/S09P31B203!84

+++ backend/src/main/java/com/ssafy/dingdong/domain/room/service/RoomServiceImpl.java
@@ -82,7 +82,7 @@ public class RoomServiceImpl implements RoomService {
 		if (category == 0) {
 			return furnitureRepository.findAllFurnitureSummaryDto(pageable);
 		}else {
-			return furnitureRepository.findAllByCategoryId(category, pageable);
+			return furnitureRepository.findAllByCategoryId(category.longValue(), pageable);
 		}
 	}
